pyproject.toml
[tool.poetry]
name = "mi-led-interface"
version = "1.0.0" # Heavy alpha.
description = "Modular interface for controlling the Merkury Innovations LED Matrix Display via BLE."
authors = ["scooglat <scooglat@gmail.com>"] # Use a real email if you plan to publish
license = "MIT" # Common, permissive license
readme = "README.md"
# homepage = "https://github.com/scooglat/my-led-display-package" # Optional: your GitHub link
# repository = "https://github.com/scooglat/my-led-display-package" # Optional: your GitHub link

# Keywords help people find your package on PyPI
keywords = ["ble", "bluetooth", "led", "matrix", "hardware"] 

# Specifies the Python versions this package supports
packages = [
    { include = "mili" }
]

[tool.poetry.dependencies]
python = "^3.10"
bleak = "^2.0.0"
pillow = "^12.0.0"
filetype = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
build = "^1.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"